assignments/assignment1/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlp-coursework1"
version = "0.1.0"
description = "Hydra-driven EMNIST experimentation stack for MLP Coursework 1"
authors = [
  { name = "Coursework Team" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "hydra-core>=1.3",
  "hydra-colorlog>=1.2",
  "hydra-optuna-sweeper>=1.2",
  "wandb>=0.17",
  "matplotlib>=3.8",
  "tensorboard>=2.16",
  "numpy>=1.26",
  "pandas>=2.2",
  "tqdm>=4.66",
  "scikit-learn>=1.3",
  "dvc>=3.60",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "ipykernel",
  "pytest",
  "black",
  "ruff"
]

[tool.setuptools.packages.find]
where = ["src"]

